naver-blog-admin/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
fastapi>=0.110.0
h11==0.16.0
idna==3.10
pydantic>=2.11.0
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy>=2.0.0
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn>=0.27.0
pydantic-settings>=2.9.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
bcrypt>=4.0.0
alembic>=1.15.0
python-dotenv>=1.0.0
email-validator>=2.1.0
aiosmtplib>=3.0.0
